move commit to action.rs
