+++ components/backup/src/service.rs
@@ -99,6 +99,7 @@ mod tests {
     use external_storage::make_local_backend;
     use security::*;
     use tikv::storage::mvcc::tests::*;
+    use tikv::storage::txn::action::commit::tests::must_succeed as must_commit;
     use tikv_util::mpsc::Receiver;
     use txn_types::TimeStamp;
