+++ src/storage/mvcc/reader/reader.rs
@@ -505,6 +505,7 @@ mod tests {
 
     use crate::storage::kv::Modify;
     use crate::storage::mvcc::{MvccReader, MvccTxn};
+    use crate::storage::txn::action;
     use concurrency_manager::ConcurrencyManager;
     use engine_rocks::properties::MvccPropertiesCollectorFactory;
     use engine_rocks::raw::DB;
@@ -637,7 +638,7 @@ mod tests {
             let start_ts = start_ts.into();
             let cm = ConcurrencyManager::new(start_ts);
             let mut txn = MvccTxn::new(snap, start_ts, true, cm);
-            txn.commit(Key::from_raw(pk), commit_ts.into()).unwrap();
+            action::commit::commit(&mut txn, Key::from_raw(pk), commit_ts.into()).unwrap();
             self.write(txn.into_modifies());
         }
