+++ src/storage/mvcc/reader/scanner/backward.rs
@@ -434,6 +434,7 @@ mod tests {
     use super::*;
     use crate::storage::kv::{Engine, TestEngineBuilder};
     use crate::storage::mvcc::tests::*;
+    use crate::storage::txn::action::commit::tests::must_succeed as must_commit;
     use crate::storage::Scanner;
     use kvproto::kvrpcpb::Context;
