+++ src/storage/txn/commands/check_txn_status.rs
@@ -157,6 +157,7 @@ pub mod tests {
     use crate::storage::kv::{Engine, WriteData};
     use crate::storage::lock_manager::DummyLockManager;
     use crate::storage::mvcc::tests::*;
+    use crate::storage::txn::action::commit::tests::must_succeed as must_commit;
     use crate::storage::txn::commands::{pessimistic_rollback, WriteCommand, WriteContext};
     use crate::storage::{types::TxnStatus, ProcessResult, TestEngineBuilder};
     use concurrency_manager::ConcurrencyManager;
