+++ src/storage/txn/commands/pessimistic_rollback.rs
@@ -94,6 +94,7 @@ pub mod tests {
     use crate::storage::kv::Engine;
     use crate::storage::lock_manager::DummyLockManager;
     use crate::storage::mvcc::tests::*;
+    use crate::storage::txn::action::commit::tests::must_succeed as must_commit;
     use crate::storage::txn::commands::{WriteCommand, WriteContext};
     use crate::storage::TestEngineBuilder;
     use concurrency_manager::ConcurrencyManager;
