+++ src/storage/txn/commands/resolve_lock.rs
@@ -7,7 +7,7 @@ use crate::storage::txn::commands::{
     Command, CommandExt, ReleasedLocks, ResolveLockReadPhase, TypedCommand, WriteCommand,
     WriteContext, WriteResult,
 };
-use crate::storage::txn::{Error, ErrorInner, Result};
+use crate::storage::txn::{action, Error, ErrorInner, Result};
 use crate::storage::{ProcessResult, Snapshot};
 use tikv_util::collections::HashMap;
 use txn_types::{Key, Lock, TimeStamp};
@@ -84,7 +84,7 @@ impl<S: Snapshot, L: LockManager> WriteCommand<S, L> for ResolveLock {
             let released = if commit_ts.is_zero() {
                 txn.rollback(current_key.clone())?
             } else if commit_ts > current_lock.ts {
-                txn.commit(current_key.clone(), commit_ts)?
+                action::commit::commit(&mut txn, current_key.clone(), commit_ts)?
             } else {
                 return Err(Error::from(ErrorInner::InvalidTxnTso {
                     start_ts: current_lock.ts,
