+++ tests/benches/hierarchy/mvcc/mod.rs
@@ -6,6 +6,7 @@ use kvproto::kvrpcpb::Context;
 use test_util::KvGenerator;
 use tikv::storage::kv::{Engine, WriteData};
 use tikv::storage::mvcc::{self, MvccReader, MvccTxn};
+use tikv::storage::txn::action;
 use txn_types::{Key, Mutation, TimeStamp};
 
 use super::{BenchConfig, EngineFactory, DEFAULT_ITERATIONS, DEFAULT_KV_GENERATOR_SEED};
@@ -87,7 +88,7 @@ fn mvcc_commit<E: Engine, F: EngineFactory<E>>(b: &mut Bencher, config: &BenchCo
         |(snapshot, keys)| {
             for key in keys {
                 let mut txn = mvcc::MvccTxn::new(snapshot.clone(), 1.into(), true, cm.clone());
-                black_box(txn.commit(key, 1.into())).unwrap();
+                black_box(action::commit::commit(&mut txn, key, 1.into())).unwrap();
             }
         },
         BatchSize::SmallInput,
