+++ tests/benches/hierarchy/txn/mod.rs
@@ -9,6 +9,7 @@ use tikv::storage::mvcc::{self, MvccTxn};
 use txn_types::{Key, Mutation, TimeStamp};
 
 use super::{BenchConfig, EngineFactory, DEFAULT_ITERATIONS};
+use tikv::storage::txn::action;
 
 fn setup_prewrite<E, F>(
     engine: &E,
@@ -83,7 +84,7 @@ fn txn_commit<E: Engine, F: EngineFactory<E>>(b: &mut Bencher, config: &BenchCon
             for key in keys {
                 let snapshot = engine.snapshot(&ctx).unwrap();
                 let mut txn = mvcc::MvccTxn::new(snapshot, 1.into(), true, cm.clone());
-                txn.commit(key, 2.into()).unwrap();
+                action::commit::commit(&mut txn, key, 2.into()).unwrap();
                 let write_data = WriteData::from_modifies(txn.into_modifies());
                 black_box(engine.write(&ctx, write_data)).unwrap();
             }
